Add a title tag
The title tag is useful for a bunch of things, e.g. in the browser it would be displayed on the tab, in the history etc. (it currently always displays the full url there, which looks somewhat weird).

We're also collecting all blog titles in our aggregator so it can be published as a blogroll here: http://railsgirlssummerofcode.org/blog

Our code could be better and just display your url in that case, but it'd still look better if you'd provide a title.

You might want to improve this and, instead of just hardcoding the title tag, also include something about the current post (if looking at a blog post view).